+++ app/code/ThirdParty/BlogArticle/Model/Resolver/DataMapper/PostMapper.php
@@ -15,6 +15,7 @@ class PostMapper
     {
         $map = [
             'title' => 'setTitle',
+            'author' => 'setAuthor',
             'url_key' => 'setUrlKey',
             'content' => 'setContent',
             'excerpt' => 'setExcerpt',
@@ -54,6 +55,7 @@ class PostMapper
         return [
             'post_id' => $item->getPostId(),
             'title' => $item->getTitle(),
+            'author' => $item->getAuthor(),
             'url_key' => $item->getUrlKey(),
             'content' => $item->getContent(),
             'excerpt' => $item->getExcerpt(),