+++ app/code/ThirdParty/BlogArticle/Model/Resolver/Post.php
@@ -49,6 +49,7 @@ class Post implements ResolverInterface
         return [
             'post_id' => $item->getPostId(),
             'title' => $item->getTitle(),
+            'author' => $item->getAuthor(),
             'url_key' => $item->getUrlKey(),
             'content' => $item->getContent(),
             'excerpt' => $item->getExcerpt(),