+++ app/code/ThirdParty/BlogArticle/Model/Resolver/RelatedPosts.php
@@ -31,6 +31,7 @@ class RelatedPosts implements ResolverInterface
             $mapped[] = [
                 'post_id' => $item->getPostId(),
                 'title' => $item->getTitle(),
+                'author' => $item->getAuthor(),
                 'url_key' => $item->getUrlKey(),
                 'content' => $item->getContent(),
                 'excerpt' => $item->getExcerpt(),